Refactor: Rename copilot-service to chat-service and extract tool/manager logic
